Some refactors and renames. Removed some defunct classes and enums, renamed config_tab to settings_tab and similar, and fixes small bug with fast_vision_mode_coordinates.
