+++ src/gui/models/activity_log_widget.py
