+++ src/gui/models/collapsible_widget.py
